// monitor/backend/src/search/metrica/span/metrica.span.search.dto.ts
export interface MetricaQuantDTO {
  estampaTempo: string;
  quant: number;
  tipo?: string;
}

export interface MetricaMSReqDTO {
  estampaTempo: string;
  milissegundos: number;
}
export interface EntradaMetricaDTO {
  rota?: string;
  servico?: string;
  agrupamento?: string;
  periodo?: number;
  tipo?: string;
  nomeTipo?: boolean;
}

export interface EntradaMetricaLogDTO {
  range: number;
}